feat[project]: create correct config for personal-account-service, mentor-catalog-service, media-service

+++ mentor-catalog-service/src/main/java/ru/just/mentorcatalogservice/dto/MentorCardDto.java
@@ -12,6 +12,6 @@ public class MentorCardDto {
     private String shortAboutMe;
     private String longAboutMe;
     private List<String> specializations;
-    private Long studentsCount;
+    private Integer studentsCount;
     private String avatarUrl;
 }
